+++ pyuarm/__init__.py
@@ -1,2 +1,1 @@
 from pyuarm import *
-from version import version,__version__,VERSION