+++ pyuarm/util.py
@@ -1,5 +1,22 @@
 from tools.list_uarms import uarm_ports
 import pyuarm
+import os
+import json
+
+
+#version
+module_dir = os.path.dirname(os.path.abspath(__file__))
+version_file = os.path.join(module_dir, 'version.json')
+with open(version_file,'r') as v:
+    data = json.load(v)
+    v.close()
+
+# pyuarm version
+version = data['module_version']
+__version__ = VERSION = version
+
+# support version
+support_versions = data['support_versions']
 
 # Firmata
 
@@ -179,6 +196,17 @@ class NoUArmPortException(IOError):
 class UnkwonFirmwareException(Exception):
     """
     .. _UnkwonFirmwareException:
+    """
+    def __init__(self, error):
+        self.error = error
+
+    def __str__(self):
+        return repr(self.error)
+
+
+class UnSupportedFirmwareVersionException(Exception):
+    """
+    .. UnSupportFirmwareVersionException:
     """
     def __init__(self, error):
         self.error = error